Feature: Add fractional scaling parameters. Improve the start menu call. Fix: Mouse theme is not available in some areas.

+++ plugins/xrandr/xrandr-manager.cpp
@@ -49,7 +49,7 @@ XrandrManager::XrandrManager()
     mXrandrSetting = new QGSettings(SETTINGS_XRANDR_SCHEMAS);
     mChangeCompressor = new QTimer(this);
     QGSettings *mXsettings = new QGSettings(XSETTINGS_SCHEMA);
-    mScale = mXsettings->get(XSETTINGS_KEY_SCALING).toInt();
+    mScale = mXsettings->get(XSETTINGS_KEY_SCALING).toDouble();
     if(mXsettings)
         delete mXsettings;
